Add subsumption rules to unification
These rules allow the kind checker to unify two unrelated Foralls,
provided that both of them have a given type. This implementation is
similar to the subtyping rules for Foralls in the type checker before it
was folded into unification.

We don't perform "real" unification here because inferring kinds at the
kind-level is non-trivial. To be more specific:

   If we're attempting to unify an unkinded forall, regardless whether
   it appears in the left or right, adding a placeholder unsolved
   variable requires that it appears within the algorithmic context. The
   problem here is that unsolved context elements require a kind
   annotation, and adding another unsolved variable creates practically
   an infinite loop.

   We can assume that the kind of the forall's kind annotation is Type,
   which is a safe assumption to take considering that the inference
   rule for foralls makes it so a kind annotation checks against Type.