Introduce subsumption to the type checker
This makes unification and subsumption two different entities now, and
makes it so that the unification rule for foralls on both sides now deal
with skolemization.

Note that the type checker still deviates from the kind checker
especially with their use of the Variable/Skolem types as well as the
kinded/unkinded context elements.

In the future, though, I'd like to be able to unify these two systems
together gracefully.